+++ feed.php
@@ -28,7 +28,7 @@ include_once(PHPWG_ROOT_PATH.'include/functions_notification.inc.php');
  */
 function datetime_to_ts(string $datetime): int
 {
-    return strtotime($datetime);
+    return (int) strtotime($datetime);
 }
 
 /**
